Limit value of NK for cloud top calculation (#1387)
TYPE: bug fix

KEYWORDS: Deng shcu, limit updraft, cloud

SOURCE: Timothy W. Juliano and Pedro A. Jimenez (NCAR/RAL), Jimy Dudhia (NCAR/MMM)

DESCRIPTION OF CHANGES:
Problem:
The issue is in loop60 of the code. There are two exit statements within that loop. If either of these conditional statements are satisfied, then the loop exits and there is no bug. The problem arises if there is a clean exit of loop60. In this loop, variable NK (the index of the loop) goes from just below cloud base (KLCL-1) to #layers-1 (KLM). If there is a clean exit of the loop, then NK = #layers. This causes variable LTOP1 to have a value of #layers and therefore LTOP also equals #layers. When this happens, in line 2250 of the official WRF release [TMA = QQG(NK+1)*EMS(NK+1)], QQG goes out of bounds since QQG is defined from 1 to #layers and NK in this loop goes from 1 to LTOP which is #layers. Thus, it tries to access index NK+1 which is #layers+1.

Solution:
The solution is to limit the value of LTOP1 at the exit of loop60 to #layers-1 (KLM).

LIST OF MODIFIED FILES:
M phys/module_shcu_deng.F

TESTS CONDUCTED:
- We have run a 3km CONUS simulation with and without the bug fix. Without the bug fix, the simulation crashes as reported above. With the bug fix, the simulation integrates through without an issue.
- Jenkins tests all PASS.

RELEASE NOTE: A fix is introduced to the Deng shallow cumulus parameterization scheme. When the updraft top reached the domain top, the model would crash due to accessing out of bounds memory. The modification is to limit the vertical index of the updraft top.