Moving check of Ptau after conditional statements to prevent negative value in FARMS (#1394)
TYPE: bug fix

KEYWORDS: FARMS, radiation, cloud optical thickness

SOURCE: Timothy W. Juliano and Pedro A. Jimenez (NCAR/RAL)

DESCRIPTION OF CHANGES:
Problem:
In the ICEMODEL subroutine, there is an issue with the calculation of Ptau (lines 602-608 in the official WRF version). If De <= 26.0 and solarangle (cosine of solar zenith angle) is really small, then Ptau (ice cloud optical thickness) can be negative. In my case, Ptau goes negative and causes the model to crash since PPDHI is later calculated using log10(Ptau). It looks like the original code accounted for a small value of Ptau but only when De > 26.0.

Solution:
I fixed the issue simply by moving the check on Ptau to be located after the conditional statements.

LIST OF MODIFIED FILES:
M phys/module_ra_farms.F

TESTS CONDUCTED:
- We have run a 3km CONUS simulation with and without the bug fix. Without the bug fix, the simulation crashes as reported above. With the bug fix, the simulation integrates through without an issue.
- Jenkins tests are passing.

RELEASE NOTE: A fix is introduced to the FARMS parameterization. When the cosine of solar zenith angle is small, the ice cloud optical thickness can be negative, and the model can crash due to log10 of a negative number. The modification is to limit the value of cloud optical thickness.