bugfix Mp38 allocatable array dimension (#1834)
Large memory requirement in Thompson microphysics mp=8 and mp=28

TYPE: bug fix

KEYWORDS: memory, Thompson, microphysics

SOURCE: Maria Frediani (NCAR)

DESCRIPTION OF CHANGES:
The new graupel (qg) table created for mp=38 required a larger memory to either compute or use (PR#1808). Because of that, when using options 8 or 28, the model stopped right away if there wasn't enough memory.

These changes make the 5D arrays truly allocatable, meaning when Thompson option 8 or 28 are used, the middle dimension (NRHG) is 1, and when 38 is used that dimension becomes 9. This should reduce the memory required when option 8 and 28 are used, and the newer qg table written out by the new code won't be as huge as it is now, but more similar to what it was before option 38 was introduced.

Also, the table was initially named qr_acr_qg_mp28V4, but this table is also used for option 8. For this reason, the name was changed to qr_acr_qg_V4.

LIST OF MODIFIED FILES: phys/module_mp_thompson.F

TESTS CONDUCTED:

1. Bit for bit comparison using diffwrf to compare the results between the current develop branch and the new code using mp=38 and mp=28.
2. It passes the regression tests.